upload

+++ scenes/gameplay/World.gd
@@ -44,21 +44,24 @@ func _input(event):
 
 # process stuff
 func _on_tile_set_request(position: Vector2, clear: bool):
-	if clear:
+	if clear and EntityManager.EntityAt(position):
 		EntityManager.RemoveEntity(position).Delete()
 
 	elif not EntityManager.EntityAt(position):
-		var belt = EntityManager.CreateNewEntity("Entity", {})
+		var belt = EntityManager.CreateNewEntity("BeltEntity", {})
 
 		var rect = RectangleShape2D.new()
 		rect.extents = Vector2(16, 14)
 		
-		belt.ConfigAnimatedEntity("res://assets/sprites/entities/belt-1/belt-1-rf.png", [2, 2], 1.5, true)
+		belt.ConfigAnimatedEntity("res://assets/sprites/entities/belt-1/belt-1-{0}f.png", [2, 2], 1.5, true)
 		belt.Init($World/Ground)
 		belt.Coordinates = position
 		
 		$World/Entities/Layer1.add_child(belt)
 		EntityManager.SetEntity(belt, position)
+	
+	else:
+		EntityManager.GetEntity(position).Rotate(true);
 
 # save stuff
 func update_total_time():
